mode get_dummy_compressed_model to common
